+++ federation/tests/utils/test_activitypub.py
@@ -41,12 +41,13 @@ class TestGetProfileIdFromWebfinger:
 
 
 class TestRetrieveAndParseDocument:
-    @pytest.mark.skip
     @patch("federation.utils.activitypub.fetch_document", autospec=True, return_value=(None, None, None))
     def test_calls_fetch_document(self, mock_fetch):
         retrieve_and_parse_document("https://example.com/foobar")
+        # auth argument is passed with kwargs
+        auth = mock_fetch.call_args.kwargs.get('auth', None)
         mock_fetch.assert_called_once_with(
-            "https://example.com/foobar", extra_headers={'accept': 'application/activity+json'},
+            "https://example.com/foobar", extra_headers={'accept': 'application/activity+json'}, auth=auth,
         )
 
     @patch("federation.utils.activitypub.fetch_document", autospec=True, return_value=(